requirements.txt
ffmpeg
python-dotenv
gradio==3.38.0
openai
tiktoken
tinydb
tinymongo
proglog
yt-dlp
torch
torchaudio
### whisper timestamped 
whisper-timestamped
protobuf==3.20.0
pillow==9.0.0
moviepy==1.0.3
progress
questionary
edge-tts
flask
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
tiktok-uploader
oauth2client